some clean up of popup, placement casing and moved constants, transforms and offsets in "utils" moved to a nested utils directory and separated.
